pyproject.toml
[project]
name = "serenade"
version = "0.1.0"
description = ""
authors = [
    {name = "redacted"}
]
readme = "README.md"
requires-python =  ">=3.14"
dependencies = [
    "discord-py[voice] (>=2.6.4,<3.0.0)",
    "yt-dlp (>=2025.12.8,<2026.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ytmusicapi (>=1.11.5,<2.0.0)",
    "spotifyscraper (>=2.1.5,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
